CSS Markers freeCodeCamp Project
Completed the next lesson in the responsive web design course. This time I decided to mess around quite a bit with the CSS for the markers. Getting the divs centered was quite the challenge but "display: flex;" and "justify-content: center;" were a god send. Definitely will not be forgetting to use that in the future for making other sites.