HTML for the Accessability Quiz
This one was very tedious honestly, but I'm dedicated to actually learning how to do everything from the ground up, because if I don't understand how everything actually works and why it is used the way it is, I won't be good at the shortcuts later. I want a very solid foundation to build upon later so I the later bits are able to be learned much faster.