HTML, and scripts started
I know that having a data file is never actually going to be how I do things IRL, but I needed a way to practice modules, and needed a way to work on manipulating data structures so I'm using a "data.js" file to simulate that. Later I will be using APIs or Databases to get the data that I will be using for the applications but needed a way to work it into the process.