Rewrote scipts for webpage
Added functionality to the website and literally spent all day trying to get the modules to work, ended up having to install node and start a localhost server to get around the CORS stuff. This was a major pain and overlooked a lot of small details with naming variables and order of the for loops. Massive headache and ready to move on honestly.