Finished Styling the webpage
Added all the CSS needed to finish the site and for it to not look like absolute garbage. Kinda went boring for the colors, but being color blind it felt like the right decision since being accessable was kinda the whole point. I feel like colorblindness isn't really thought of that much since it's not something complete life changing.