Reworked HTML for blog
Am going to add some functionality to this so that I can update it as I go along. I want to make a button to create a new post, edit previous posts, a search function and the ability to remove posts, I want to keep track of those posts I deleted too so I'm definitely adding a restore function too. We will see what happens!