Changed code; using modules
This is going to be the main script that sets up the page and actually executes the methods from my app object. This kinda just feels like the best way to do things. I think I'm doing OOP but could be completely wrong. I am having a lot of fun figuring all this out though!